Issue/30 feature item crud

+++ src/main/java/com/example/place/domain/item/service/ItemService.java
@@ -1,7 +1,7 @@
 package com.example.place.domain.item.service;
 
-import com.example.place.domain.item.dto.request.CreateItemRequest;
-import com.example.place.domain.item.dto.response.CreateItemResponse;
+import com.example.place.domain.item.dto.request.ItemRequest;
+import com.example.place.domain.item.dto.response.ItemResponse;
 import com.example.place.domain.item.entity.Item;
 import com.example.place.domain.item.repository.ItemRepository;
 import com.example.place.domain.itemtag.entity.ItemTag;
@@ -21,6 +21,9 @@ import com.example.place.common.exception.exceptionclass.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
+import java.util.List;
+
 
 @Service
 @RequiredArgsConstructor
@@ -31,11 +34,6 @@ public class ItemService {
     private final TagRepository tagRepository;
     private final UserRepository userRepository;
 
-	public Item findItemById(Long itemId){
-		Item item = itemRepository.findById(itemId)
-			.orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_ITEM));
-		return item;
-	}
 
 	// 재고 감소
 	@Transactional
@@ -55,18 +53,27 @@ public class ItemService {
 
 	//기본적인 태그 관리 흐름입니다
     @Transactional
-    public CreateItemResponse createItem(Long userId,CreateItemRequest request) {
+    public ItemResponse createItem(Long userId, ItemRequest request) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.NOT_FOUND_USER));
 
+		LocalDateTime salesStartAt = request.getSalesStartAt() != null
+				? request.getSalesStartAt()
+				: LocalDateTime.now();
+
+		LocalDateTime salesEndAt = request.getSalesEndAt() != null
+				? request.getSalesEndAt()
+				: LocalDateTime.of(3000, 1, 1, 0, 0);
+//		:LocalDateTime.Max;
         Item item = new Item(
                 user,
 				request.getItemName(),
 				request.getItemDescription(),
 				request.getPrice(),
 				request.getCount(),
-				request.getSalesStartAt(),
-				request.getSalesEndAt());
+				salesStartAt,
+				salesEndAt
+				);
         itemRepository.save(item);
 
         for (String tagName: request.getItemTagNames()) {
@@ -76,29 +83,33 @@ public class ItemService {
             ItemTag itemTag = new ItemTag(null, item, tag);
             itemTagRepository.save(itemTag);
         }
-        return CreateItemResponse.from(item);
+        return ItemResponse.from(item);
     }
 
 	@Transactional(readOnly = true)
-	public CreateItemResponse readItem(Long itemId) {
+	public ItemResponse readItem(Long itemId) {
 		Item item = findByIdOrElseThrow(itemId);
-		return CreateItemResponse.from(item);
+		return ItemResponse.from(item);
 	}
 
 	@Transactional
-	public CreateItemResponse updateItem(Long itemId, CreateItemRequest request) {
+	public ItemResponse updateItem(Long itemId, ItemRequest request, Long userId) {
 		Item item = findByIdOrElseThrow(itemId);
+
+		if(!item.getUser().getId().equals(userId)) {
+			throw new CustomException(ExceptionCode.FORBIDDEN_ITEM_ACCESS);
+		}
 		item.updateItem(request);
-		return CreateItemResponse.from(item);
+		return ItemResponse.from(item);
 	}
 
 	@Transactional
 	public void deleteItem(Long itemId, Long userId) {
-		itemRepository.deleteById(itemId);
 
-//		if(!item.getUser.getId().equals(userId)) {
-//			throw new CustomException(ExceptionCode.FORBIDDEN_ITEM_DELETE);
-//		}
+		if(!findByIdOrElseThrow(itemId).getUser().getId().equals(userId)) {
+			throw new CustomException(ExceptionCode.FORBIDDEN_ITEM_DELETE);
+		}
+		itemRepository.deleteById(itemId);
 	}
 
 
@@ -108,4 +119,10 @@ public class ItemService {
 	}
 
 
+	public List<ItemResponse> searchItem(String keyword, List<String> tags, Long userId) {
+		return itemRepository.searchItems(keyword, tags, userId)
+				.stream()
+				.map(ItemResponse::from)
+				.toList();
+	}
 }
